+++ kazoo/client.py
@@ -387,6 +387,7 @@ class KazooClient(object):
         :param default_acl: A default ACL used on node creation.
         """
         from kazoo.recipe.barrier import Barrier
+        from kazoo.recipe.barrier import DoubleBarrier
         from kazoo.recipe.partitioner import SetPartitioner
         from kazoo.recipe.watchers import ChildrenWatch
         from kazoo.recipe.watchers import DataWatch
@@ -441,6 +442,7 @@ class KazooClient(object):
 
         # convenience API
         self.Barrier = partial(Barrier, self)
+        self.DoubleBarrier = partial(DoubleBarrier, self)
         self.ChildrenWatch = partial(ChildrenWatch, self)
         self.DataWatch = partial(DataWatch, self)
         self.Election = partial(Election, self)