+++ kazoo/tests/test_barrier.py
@@ -1,3 +1,5 @@
+import threading
+
 from nose.tools import eq_
 
 from kazoo.testing import KazooTestCase
@@ -18,3 +20,115 @@ class KazooBarrierTests(KazooTestCase):
     def test_remove_nonexistent_barrier(self):
         b = self.client.Barrier("/some/path")
         eq_(b.remove(), False)
+
+
+class KazooDoubleBarrierTests(KazooTestCase):
+    def test_basic_barrier(self):
+        b = self.client.DoubleBarrier("/some/path", 1)
+        eq_(b.participating, False)
+        b.enter()
+        eq_(b.participating, True)
+        b.leave()
+        eq_(b.participating, False)
+
+    def test_two_barrier(self):
+        av = threading.Event()
+        ev = threading.Event()
+        bv = threading.Event()
+        release_all = threading.Event()
+        b1 = self.client.DoubleBarrier("/some/path", 2)
+        b2 = self.client.DoubleBarrier("/some/path", 2)
+
+        def make_barrier_one():
+            b1.enter()
+            ev.set()
+            release_all.wait()
+            b1.leave()
+            ev.set()
+
+        def make_barrier_two():
+            bv.wait()
+            b2.enter()
+            av.set()
+            release_all.wait()
+            b2.leave()
+            av.set()
+
+        # Spin up both of them
+        t1 = threading.Thread(target=make_barrier_one)
+        t1.start()
+        t2 = threading.Thread(target=make_barrier_two)
+        t2.start()
+
+        eq_(b1.participating, False)
+        eq_(b2.participating, False)
+
+        bv.set()
+        av.wait()
+        ev.wait()
+        eq_(b1.participating, True)
+        eq_(b2.participating, True)
+
+        av.clear()
+        ev.clear()
+
+        release_all.set()
+        av.wait()
+        ev.wait()
+        eq_(b1.participating, False)
+        eq_(b2.participating, False)
+
+    def test_three_barrier(self):
+        av = threading.Event()
+        ev = threading.Event()
+        bv = threading.Event()
+        release_all = threading.Event()
+        b1 = self.client.DoubleBarrier("/some/path", 3)
+        b2 = self.client.DoubleBarrier("/some/path", 3)
+        b3 = self.client.DoubleBarrier("/some/path", 3)
+
+        def make_barrier_one():
+            b1.enter()
+            ev.set()
+            release_all.wait()
+            b1.leave()
+            ev.set()
+
+        def make_barrier_two():
+            bv.wait()
+            b2.enter()
+            av.set()
+            release_all.wait()
+            b2.leave()
+            av.set()
+
+        # Spin up both of them
+        t1 = threading.Thread(target=make_barrier_one)
+        t1.start()
+        t2 = threading.Thread(target=make_barrier_two)
+        t2.start()
+
+        eq_(b1.participating, False)
+        eq_(b2.participating, False)
+
+        bv.set()
+        eq_(b1.participating, False)
+        eq_(b2.participating, False)
+        b3.enter()
+        ev.wait()
+        av.wait()
+
+        eq_(b1.participating, True)
+        eq_(b2.participating, True)
+        eq_(b3.participating, True)
+
+        av.clear()
+        ev.clear()
+
+        release_all.set()
+        b3.leave()
+        av.wait()
+        ev.wait()
+        eq_(b1.participating, False)
+        eq_(b2.participating, False)
+        eq_(b3.participating, False)